Bootsrap react-native
