+++ __tests__/index.android.js
@@ -1,9 +1,9 @@
 import 'react-native';
-import * as React from 'react';
-import Index from '../index.ios';
+import React from 'react';
+import Index from '../index.android.js';
 
 // Note: test renderer must be required after react-native.
-import * as renderer from 'react-test-renderer';
+import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
   const tree = renderer.create(
