+++ __tests__/index.ios.js
@@ -1,9 +1,9 @@
 import 'react-native';
-import * as React from 'react';
-import Index from '../index.android';
+import React from 'react';
+import Index from '../index.ios.js';
 
 // Note: test renderer must be required after react-native.
-import * as renderer from 'react-test-renderer';
+import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
   const tree = renderer.create(
