+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.myawesomeproject",
+    package = "com.test",
 )
 
 android_resource(
     name = "res",
-    package = "com.myawesomeproject",
+    package = "com.test",
     res = "src/main/res",
 )
