+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'MyAwesomeProject'
+rootProject.name = 'Test'
 
 include ':app'
