+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "MyAwesomeProject",
-  "displayName": "MyAwesomeProject"
+  "name": "Test",
+  "displayName": "Test"
 }
