+++ ios/Test.xcodeproj/xcshareddata/xcschemes/Test-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "MyAwesomeProject-tvOS.app"
-               BlueprintName = "MyAwesomeProject-tvOS"
-               ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+               BuildableName = "Test-tvOS.app"
+               BlueprintName = "Test-tvOS"
+               ReferencedContainer = "container:Test.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "MyAwesomeProject-tvOSTests.xctest"
-               BlueprintName = "MyAwesomeProject-tvOSTests"
-               ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+               BuildableName = "Test-tvOSTests.xctest"
+               BlueprintName = "Test-tvOSTests"
+               ReferencedContainer = "container:Test.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "MyAwesomeProject-tvOSTests.xctest"
-               BlueprintName = "MyAwesomeProject-tvOSTests"
-               ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+               BuildableName = "Test-tvOSTests.xctest"
+               BlueprintName = "Test-tvOSTests"
+               ReferencedContainer = "container:Test.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "MyAwesomeProject-tvOS.app"
-            BlueprintName = "MyAwesomeProject-tvOS"
-            ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+            BuildableName = "Test-tvOS.app"
+            BlueprintName = "Test-tvOS"
+            ReferencedContainer = "container:Test.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "MyAwesomeProject-tvOS.app"
-            BlueprintName = "MyAwesomeProject-tvOS"
-            ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+            BuildableName = "Test-tvOS.app"
+            BlueprintName = "Test-tvOS"
+            ReferencedContainer = "container:Test.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "MyAwesomeProject-tvOS.app"
-            BlueprintName = "MyAwesomeProject-tvOS"
-            ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+            BuildableName = "Test-tvOS.app"
+            BlueprintName = "Test-tvOS"
+            ReferencedContainer = "container:Test.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
