+++ ios/Test.xcodeproj/xcshareddata/xcschemes/Test.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "MyAwesomeProject.app"
-               BlueprintName = "MyAwesomeProject"
-               ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+               BuildableName = "Test.app"
+               BlueprintName = "Test"
+               ReferencedContainer = "container:Test.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "MyAwesomeProjectTests.xctest"
-               BlueprintName = "MyAwesomeProjectTests"
-               ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+               BuildableName = "TestTests.xctest"
+               BlueprintName = "TestTests"
+               ReferencedContainer = "container:Test.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "MyAwesomeProjectTests.xctest"
-               BlueprintName = "MyAwesomeProjectTests"
-               ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+               BuildableName = "TestTests.xctest"
+               BlueprintName = "TestTests"
+               ReferencedContainer = "container:Test.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "MyAwesomeProject.app"
-            BlueprintName = "MyAwesomeProject"
-            ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+            BuildableName = "Test.app"
+            BlueprintName = "Test"
+            ReferencedContainer = "container:Test.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "MyAwesomeProject.app"
-            BlueprintName = "MyAwesomeProject"
-            ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+            BuildableName = "Test.app"
+            BlueprintName = "Test"
+            ReferencedContainer = "container:Test.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "MyAwesomeProject.app"
-            BlueprintName = "MyAwesomeProject"
-            ReferencedContainer = "container:MyAwesomeProject.xcodeproj">
+            BuildableName = "Test.app"
+            BlueprintName = "Test"
+            ReferencedContainer = "container:Test.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
