+++ ios/Test/Images.xcassets/AppIcon.appiconset/Contents.json
